handles the bulk of the uploading in the store.js file now instead of in the actions
