+++ package.json
@@ -38,9 +38,11 @@
     "enzyme": "^3.8.0",
     "enzyme-adapter-react-16": "^1.9.1",
     "enzyme-to-json": "^3.3.5",
+    "firebase-functions-test": "^0.1.6",
     "react-scripts": "^2.1.5",
     "react-test-renderer": "^16.8.1",
-    "redux-devtools-extension": "^2.13.8"
+    "redux-devtools-extension": "^2.13.8",
+    "redux-mock-store": "^1.5.3"
   },
   "scripts": {
     "start": "react-scripts start",
